refactor(pci-instances): separate mapper from selector and make test more relevant
