+++ packages/manager/apps/pci-instances/src/data/api/instance.spec.ts
@@ -2,7 +2,6 @@ import { describe, it, vi } from 'vitest';
 import { v6 } from '@ovh-ux/manager-core-api';
 import { editInstanceName } from './instance';
 
-const put = vi.fn();
 const projectId = 'projectId-test';
 const instanceId = 'instanceId-test';
 const instanceName = 'instanceName-test';
@@ -11,11 +10,10 @@ vi.mock('@ovh-ux/manager-core-api');
 
 describe('editInstanceName', () => {
   it('should call the good endpoint with the instance name params body', () => {
-    vi.spyOn(v6, 'put').mockImplementation(put);
     editInstanceName({ projectId, instanceId, instanceName });
 
     expect(
-      put,
+      v6.put,
     ).toHaveBeenCalledWith(
       `/cloud/project/${projectId}/instance/${instanceId}`,
       { instanceName },
