+++ packages/manager/apps/pci-instances/src/pages/instances/instance/Instance.page.tsx
@@ -1,4 +1,4 @@
-import { FC, useMemo } from 'react';
+import { FC } from 'react';
 import { useParams, useRouteLoaderData } from 'react-router-dom';
 import { OsdsSkeleton } from '@ovhcloud/ods-components/react';
 import { ODS_SKELETON_SIZE } from '@ovhcloud/ods-components';
@@ -16,19 +16,16 @@ import { Breadcrumb } from '@/components/breadcrumb/Breadcrumb.component';
 import { CHANGELOG_LINKS } from '@/constants';
 import { useInstance } from '@/data/hooks/instance/useInstance';
 import InstanceName from './component/InstanceName.component';
-import { instanceSelector } from '@/data/hooks/instance/selectors/instances.selector';
+import { mapInstanceDto } from '@/data/hooks/instance/mapper/instance.mapper';
 
 const Instance: FC = () => {
   const project = useRouteLoaderData('root') as TProject;
   const { instanceId } = useParams() as { instanceId: string };
   const projectUrl = useProjectUrl('public-cloud');
 
-  const { data, isLoading } = useInstance(instanceId);
-
-  const instance = useMemo(
-    () => (data ? instanceSelector(data, projectUrl) : null),
-    [data],
-  );
+  const { data: instance, isLoading } = useInstance(instanceId, {
+    select: (dto) => mapInstanceDto(dto, projectUrl),
+  });
 
   return (
     <InstanceWrapper>
@@ -43,7 +40,13 @@ const Instance: FC = () => {
           <div className="flex items-center justify-between">
             <div className="flex-[0.8]">
               {isLoading && <OsdsSkeleton size={ODS_SKELETON_SIZE.sm} inline />}
-              {instance && <InstanceName instance={instance} />}
+              {instance && (
+                <InstanceName
+                  instanceId={instance.id}
+                  status={instance.status.label}
+                  name={instance.name}
+                />
+              )}
             </div>
             <div className="flex gap-x-3">
               <ChangelogButton links={CHANGELOG_LINKS} />
