+++ packages/manager/apps/pci-instances/src/pages/instances/instance/component/InstanceName.spec.tsx
@@ -2,18 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { describe, it, vi, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { useNotifications } from '@ovh-ux/manager-react-components';
-import {
-  instanceBuilder,
-  instanceDtoBuilder,
-} from '@/__mocks__/instance/builder';
 import InstanceName from './InstanceName.component';
-import { TInstance } from '@/types/instance/entity.type';
-
-const fakeInstanceDtoWithError = instanceDtoBuilder([], 'ERROR');
-const fakeInstance = instanceBuilder(fakeInstanceDtoWithError, new Map(), {
-  label: 'ERROR',
-  severity: 'error',
-});
 
 vi.mock('@/data/api/instance');
 
@@ -21,16 +10,17 @@ vi.mocked(useNotifications).mockReturnValue({
   addError: vi.fn(),
 });
 
-const renderInstanceName = ({ instance }: { instance: TInstance }) =>
-  render(
-    <QueryClientProvider client={new QueryClient()}>
-      <InstanceName instance={instance} />
-    </QueryClientProvider>,
-  );
-
 describe('Instance Name', () => {
-  it('Should not editable if instance is on Error', () => {
-    renderInstanceName({ instance: fakeInstance });
+  it('Should not be editable if instance is in Error state', () => {
+    render(
+      <QueryClientProvider client={new QueryClient()}>
+        <InstanceName
+          instanceId="fake-id"
+          status="ERROR"
+          name="fake-instance-name"
+        />
+      </QueryClientProvider>,
+    );
 
     expect(screen.getByText('fake-instance-name')).toBeInTheDocument();
     expect(screen.queryByTestId('edit-btn')).not.toBeInTheDocument();
